+++ android/project/app/src/main/java/com/chill/mallang/ui/feature/incorrect_word/IncorrectWord.kt
@@ -1,6 +1,6 @@
 package com.chill.mallang.ui.feature.incorrect_word
 
 data class IncorrectWord(
-    val studyId: Int,
+    val studyId: Long,
     val script: String, // 문제 script
 )
